pyproject.toml
[project]
name = "psai-bench"
version = "1.0.0rc1"
description = "PSAI-Bench: Physical Security AI Triage Benchmark"
readme = {text = "PSAI-Bench: Physical Security AI Triage Benchmark", content-type = "text/plain"}
license = "Apache-2.0"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "jsonschema>=4.0",
    "click>=8.0",
    "tabulate>=0.9",
    "matplotlib>=3.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
]
api = [
    "anthropic>=0.40",
    "openai>=1.0",
    "google-genai>=1.0",
]

[project.scripts]
psai-bench = "psai_bench.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["psai_bench*"]

[tool.ruff]
line-length = 100
